Outline for the isolated lwt_interface code.
