+++ python/setup.py
@@ -54,7 +54,7 @@ _tskit_module = Extension(
     extra_compile_args=["-std=c99"],
     libraries=libraries,
     define_macros=defines,
-    include_dirs=[libdir, kastore_dir],
+    include_dirs=["lwt_interface", libdir, kastore_dir],
 )
 
 here = os.path.abspath(os.path.dirname(__file__))
